refactor 프로젝트 패키지, 클래스 구조 개선
- 웹만을 고려하지 않고 확장성 있는 의미로 web 패키지 -> api 패키지로 명칭 변경
- 영속성 관리 차원에서 persistence 패키지 내부에 rdbms, searchengine 패키지 이동
- logs 패키지는 controller에서 활용되는 부분이므로 api 패키지 하위로 이동
- security 관련 설정은 전 프로젝트를 아우르므로 common 패키지 하위에 security 패키지 생성 후 관리 예정
- 실제 외부 영속성과 통신하는 부분은 Repository로 끝나도록 명칭 변경 (예외는 존재)
- 가독성 향상을 위해 adapter layer에서 명시적으로 entity 패키지를 생성하여 관리
- domain 패키지는 비즈니스 로직을 구현하는 부분인 application 패키지 내부로
이동하는게 더 자연스럽다는 의견 반영
