--- webtoon-search-back/webtoon-search/Dockerfile-spring-original
@@ -1,16 +1,0 @@
-FROM amazoncorretto:21-alpine-jdk
-
-WORKDIR /imageTest
-
-# 프로젝트 파일 복사
-COPY . .
-
-# gradlew에 실행 권한 부여
-RUN chmod +x ./gradlew
-
-# Gradle 빌드 실행 (셸을 명시적으로 지정)
-RUN /bin/sh -c "./gradlew build --no-daemon"
-
-# JAR 파일 실행
-EXPOSE 8080
-CMD ["java", "-jar", "build/libs/webtoon-search-*.jar"]
