+++ webtoon-search-back/webtoon-search/src/main/java/com/samsamohoh/webtoonsearch/adapter/api/member/OAuth2UserInfo.java
@@ -1,0 +1,16 @@
+package com.samsamohoh.webtoonsearch.adapter.api.member;
+
+import java.util.Map;
+
+public interface  OAuth2UserInfo {
+
+    Map<String, Object> getAttributes();
+    String getProviderId();
+    String getEmail();
+    String getAge();
+    String getGender();
+    String getNickname();
+    String getProvider();
+    String getName();
+
+}
