--- webtoon-search-back/webtoon-search/src/main/java/com/samsamohoh/webtoonsearch/adapter/persistence/MemberPersistenceAdapter.java
@@ -1,7 +1,0 @@
-package com.samsamohoh.webtoonsearch.adapter.persistence;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberPersistenceAdapter extends JpaRepository<RegisterMemberEntity, Long> {
-    RegisterMemberEntity findByProviderId(String providerId);
-}
