--- webtoon-search-back/webtoon-search/src/main/java/com/samsamohoh/webtoonsearch/adapter/persistence/SearchPersistenceAdpater.java
@@ -1,6 +1,0 @@
-package com.samsamohoh.webtoonsearch.adapter.persistence;
-
-public interface SearchPersistenceAdpater {
-
-
-}
