--- webtoon-search-back/webtoon-search/src/main/java/com/samsamohoh/webtoonsearch/adapter/persistence/SelectPersistenceAdapter.java
@@ -1,7 +1,0 @@
-package com.samsamohoh.webtoonsearch.adapter.persistence;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SelectPersistenceAdapter extends JpaRepository<SelectRecordEntity, String> {
-
-}
