+++ webtoon-search-back/webtoon-search/src/main/java/com/samsamohoh/webtoonsearch/adapter/persistence/rdbms/SelectJpaRepository.java
@@ -1,0 +1,8 @@
+package com.samsamohoh.webtoonsearch.adapter.persistence.rdbms;
+
+import com.samsamohoh.webtoonsearch.adapter.persistence.rdbms.entity.SelectRecordEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SelectJpaRepository extends JpaRepository<SelectRecordEntity, String> {
+
+}
