+++ webtoon-search-back/webtoon-search/src/main/java/com/samsamohoh/webtoonsearch/adapter/persistence/searchengine/SearchWebtoonEntityEdit.java
