--- webtoon-search-back/webtoon-search/src/main/java/com/samsamohoh/webtoonsearch/adapter/web/member/OAuthController.java
@@ -1,33 +1,0 @@
-package com.samsamohoh.webtoonsearch.adapter.web.member;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.Map;
-
-
-@RestController
-@RequestMapping("/logins")
-@RequiredArgsConstructor
-public class OAuthController {
-
-
-    @GetMapping("/loginInfo")
-    @ResponseBody
-    public Map<String, Object>  oauthLoginInfo(Authentication authentication, @AuthenticationPrincipal OAuth2User oAuth2UserPrincipal) {
-
-        OAuth2User oAuth2User = (OAuth2User) authentication.getPrincipal();
-        Map<String, Object> attributes = oAuth2User.getAttributes();
-
-        System.out.println("attributes::" + attributes);
-
-        return attributes;
-    }
-
-}
