--- webtoon-search-back/webtoon-search/src/main/java/com/samsamohoh/webtoonsearch/domain/MemberValidation.java
@@ -1,17 +1,0 @@
-package com.samsamohoh.webtoonsearch.domain;
-
-import lombok.Builder;
-import lombok.Value;
-
-@Value
-@Builder
-public class MemberValidation {
-    String email;
-    String gender;
-    String providerId;
-    String age;
-    String nickname;
-    String provider;
-
-
-}
